backend/my-app-name/new_app/serializers.py
from rest_framework import serializers
from new_app.models import ExampleModel

class ExampleModelSerializer(serializers.ModelSerializer):
    class Meta:
        model = ExampleModel
        fields = ('firstname', 'lastname')